Dodavanje Mooda i Statistika
-Mood se dodaje u bazu podataka
-Promjenjen tip podataka za MoodValue sa int na double
-Odradena Statistika tacnije graf i potpuno funkcionalan na frontendu i bekendu.

+++ backend/Dtos/MoodDto.cs
@@ -2,7 +2,7 @@
 {
     public class MoodDto
     {
-        public int MoodValue { get; set; }
+        public double MoodValue { get; set; }
     }
 }
